requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-wtf
requests
python-dotenv
gunicorn
flask-login